+++ client/scripts/option-menu.js
@@ -1,7 +1,7 @@
 import * as basic from "../scripts/basics.js";
-import * as search from "../scripts/search.js";
 import * as notify from "../scripts/notify.js";
-import * as folderData from "../scripts/folder-data.js";
+import * as optionTitle from "../scripts/option-title.js";
+import * as optionDelete from "../scripts/option-delete.js";
 import * as showAvailableVideos from "../scripts/show-available-videos.js";
 import * as currentVideoDownloads from "../scripts/current-video-downloads.js";
 
@@ -53,7 +53,7 @@ export function optionVideoMenuOnClick(videoSrc, videoType, videoInfo_ID, video_
           if (keyCode == "Enter"){ 
             if (video_name !== inputNewTitle.value) {
               video_name = inputNewTitle.value;
-              showAvailableVideos.changeVideoTitle(videoInfo_ID, video_name);
+              optionTitle.changeVideoTitle(videoInfo_ID, video_name);
             }
             inputNewTitle.blur();
             return false;
@@ -88,7 +88,7 @@ export function optionVideoMenuOnClick(videoSrc, videoType, videoInfo_ID, video_
                     if (keyCode == "Enter"){
                       if (video_name !== inputNewTitle.value) {
                         video_name = inputNewTitle.value;
-                        showAvailableVideos.changeVideoTitle(videoInfo_ID, video_name);   
+                        optionTitle.changeVideoTitle(videoInfo_ID, video_name);   
                       }
                       if (hovered  === false) {
                         document.getElementById(`${videoInfo_ID}-title`).remove();
@@ -163,7 +163,7 @@ export function  optionFolderMenuOnClick(savedVideosThumbnailContainer, folderIn
       if (keyCode == "Enter"){ 
         if (folder_name !== inputNewTitle.value) {
           folder_name = inputNewTitle.value;
-          showAvailableVideos.changeVideoTitle(folderInfo_ID, folder_name); 
+          optionTitle.changeVideoTitle(folderInfo_ID, folder_name); 
         }
         inputNewTitle.blur();
         return false;
@@ -198,7 +198,7 @@ export function  optionFolderMenuOnClick(savedVideosThumbnailContainer, folderIn
                 if (keyCode == "Enter"){
                   if (folder_name !== inputNewTitle.value) {
                     folder_name = inputNewTitle.value;
-                    showAvailableVideos.changeVideoTitle(folderInfo_ID, folder_name);   
+                    optionTitle.changeVideoTitle(folderInfo_ID, folder_name);   
                   }   
                   if (hovered  === false) {
                     if (document.getElementById(`${folderInfo_ID}-title`)) {  
@@ -310,7 +310,7 @@ export function optionVideoMenuEditOnClick(videoSrc, videoType, videoInfo_ID, vi
         }
         if (video_name !== inputNewTitle.value) {
           video_name = inputNewTitle.value;
-          showAvailableVideos.changeVideoTitle(videoInfo_ID, video_name); 
+          optionTitle.changeVideoTitle(videoInfo_ID, video_name); 
         }
         linkContainer.href = `${window.location.origin}/?t=${videoType}?v=${window.location.origin}${videoSrc}`;
         option_menu.classList = "thumbnail-option-menu fa fa-bars";
@@ -345,7 +345,7 @@ export function optionVideoMenuEditOnClick(videoSrc, videoType, videoInfo_ID, vi
           video_edit_container.remove();
           if (document.getElementById(`${videoInfo_ID}-title`)) { 
             video_name = video_title_edit_content_input.value;
-            showAvailableVideos.changeVideoTitle(videoInfo_ID, video_name); 
+            optionTitle.changeVideoTitle(videoInfo_ID, video_name); 
           } else {
             notify.message("error",`ID ${videoInfo_ID}-title is Missing`); 
           }
@@ -376,7 +376,7 @@ export function optionVideoMenuEditOnClick(videoSrc, videoType, videoInfo_ID, vi
             document.body.style.removeProperty("overflow");
             video_edit_container.remove();
             //delete data permanently
-            deleteVideoDataPermanently(videoInfo_ID);
+            optionDelete.deleteVideoDataPermanently(videoInfo_ID);
           };
         };
         return "optionVideoMenuEditOnClick";
@@ -394,7 +394,7 @@ export function optionFolderMenuEditOnClick(folderInfo_ID, folder_name, option_m
     }
     if (folder_name !== inputNewTitle.value) {
         folder_name = inputNewTitle.value;
-        showAvailableVideos.changeVideoTitle(folderInfo_ID, folder_name); 
+        optionTitle.changeVideoTitle(folderInfo_ID, folder_name); 
     }
     option_menu.classList = "thumbnail-option-menu fa fa-bars";
     option_menu_container.remove();
@@ -428,7 +428,7 @@ export function optionFolderMenuEditOnClick(folderInfo_ID, folder_name, option_m
         folder_edit_container.remove();
         if (document.getElementById(`${folderInfo_ID}-title`)) { 
           folder_name = folder_title_edit_content_input.value;
-          showAvailableVideos.changeVideoTitle(folderInfo_ID, folder_name); 
+          optionTitle.changeVideoTitle(folderInfo_ID, folder_name); 
         } else {
           notify.message("error",`ID ${folderInfo_ID}-title is Missing`); 
         }
@@ -459,7 +459,7 @@ export function optionFolderMenuEditOnClick(folderInfo_ID, folder_name, option_m
         document.body.style.removeProperty("overflow");
         folder_edit_container.remove();
         //delete data permanently
-        deleteVideoDataPermanently(folderInfo_ID);
+        optionDelete.deleteVideoDataPermanently(folderInfo_ID);
       };
     };
 }
@@ -490,7 +490,7 @@ export function closeVideoOptionMenuOnClick(videoSrc, videoType, videoInfo_ID, v
       } else {  
         if (video_name !== inputNewTitle.value) {
           video_name = inputNewTitle.value;
-          showAvailableVideos.changeVideoTitle(videoInfo_ID, video_name); 
+          optionTitle.changeVideoTitle(videoInfo_ID, video_name); 
         }
         document.getElementById(`${videoInfo_ID}-title`).remove();
         document.getElementById(`${videoInfo_ID}-title-container`).href = `${window.location.origin}/?t=${videoType}?v=${window.location.origin}${videoSrc}`;
@@ -513,7 +513,7 @@ export function closeVideoOptionMenuOnClick(videoSrc, videoType, videoInfo_ID, v
 export function closeFolderOptionMenuOnClick(folderInfo_ID, folder_name, option_menu, option_menu_container, close_option_menu, folderContainerLink, folderTitleContainer, inputNewTitle, folderURL) {
     if (folder_name !== inputNewTitle.value) {
       folder_name = inputNewTitle.value;
-      showAvailableVideos.changeVideoTitle(folderInfo_ID, folder_name); 
+      optionTitle.changeVideoTitle(folderInfo_ID, folder_name); 
     }
     if (document.getElementById(`${folderInfo_ID}-title`)) {  
         document.getElementById(`${folderInfo_ID}-title`).remove();
@@ -549,43 +549,5 @@ export function backToViewAvailableVideoButton(video_edit_body, video_edit_conta
       }
     } catch (error) {
       return "backToViewAvailableVideoButton didnt work";
-    }
-}
-
-// send request to server to delete video and all video data permently from the system
-export async function deleteVideoDataPermanently(videoID) {
-    try { 
-      const payload = {
-        id: videoID,
-        folderIDPath: folderData.getFolderIDPath()
-      };  
-      const response = await fetch("../delete-video-data-permanently", {
-        method: "POST",
-        headers: { "Content-Type": "application/json" },
-        body: JSON.stringify(payload),
-      }); 
-      let deleteVideoStatus;
-      if (response.ok) { 
-        deleteVideoStatus = await response.json();
-        if (deleteVideoStatus ==  `deleted-${videoID}-permanently`) {
-          //remove video from /saved/videos
-          document.getElementById(videoID).remove();
-          // delete searchable array item 
-          search.deleteIDFromSearchableVideoDataArray(videoID);
-          // display either noAvailableVideosDetails or noSearchableVideoData depending on the senario
-          search.noAvailableOrSearchableVideoMessage();
-          notify.message("success",`Deleted: ${videoID}`);
-          return `video-id-${videoID}-data-permanently-deleted`;
-        } else {
-          notify.message("error",`Failed Delete: ${videoID}`);
-          return `video-id-${videoID}-data-failed-to-permanently-deleted`;
-        }  
-      } else { 
-        notify.message("error","Failed Fetch: Video Deletion");
-        return "Failed to Complete Request";
-      } 
-    } catch (error) {  
-      notify.message("error","Failed Fetch: Video Deletion");
-      return error;
     }
 }
